feat: Add string validation functionality
This commit adds string validation functionality to the DurableFunctionExample class. It introduces an interface, IStringValidator, and two implementations, MockStringValidator and RealStringValidator. The StringValidator class is also added to validate strings using FluentValidation. The ValidationResult class is created to hold the result of the validation process.

The Startup class is modified to register the appropriate implementation of IStringValidator based on a configuration value. If "FeatureManagement:MockUnfinishedFeature" is true, it registers MockStringValidator; otherwise, it registers RealStringValidator.

These changes enable the DurableFunctionExample class to call _stringValidator.Validate(name) in order to validate the input string before processing it further.